Functionality Improvements
Added in the ability for to specify both homogeneous and disparate
colorschemes (whether or not everything shares the same saturation and
value).

Added in a printscheme function to retrieve the 256-color and hex-string
version of the palette in use. Using this will slightly speed up startup
time the next time since those colors won't need to be calculated.